Feat: Comment 추가, 삭제 구현

+++ src/main/java/controller/task/DeleteTaskController.java
@@ -10,11 +10,16 @@ public class DeleteTaskController implements Controller{
 
 	@Override
 	public String execute(HttpServletRequest request, HttpServletResponse response) throws Exception {		
-
-		TaskManager tManager = TaskManager.getInstance();
-		int taskId = Integer.parseInt(request.getParameter("taskId"));
-		System.out.println("삭제할 태스크" + taskId);
-		tManager.deleteTask(taskId);
+		if (request.getServletPath().equals("/comment/create")) {		// 추가
+		    
+		}
+		else if(request.getServletPath().equals("/comment/delete")) {	// 삭제
+			TaskManager tManager = TaskManager.getInstance();
+			int taskId = Integer.parseInt(request.getParameter("taskId"));
+			System.out.println("삭제할 태스크" + taskId);
+			tManager.deleteTask(taskId);
+		}
+		 
 		
 		return "redirect:/project/view?step=1&&projectId="+request.getParameter("projectId");
 	}
